requirements.txt
fast-dash
embedchain==0.0.30
gunicorn
python-dotenv==1.0.0